WebSockets/tests/submissions/Opera/cookies/004.html

<!doctype html>
<title>WebSockets: setting HttpOnly cookies in ws response, checking document.cookie</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>

setup({timeout:10001})

var t = async_test(function(t) {
  var iframe = document.createElement('iframe');
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/ws/set-cookie_http');
  ws.onopen = t.step_func(function(e) {
    ws.close();
    assert_false(!(/ws_test=test/.test(document.cookie)));
    t.done();
  });
  ws.onclose = t.step_func(function() {assert_unreached()});


  add_result_callback(function() {
    //This might be racy if a completion callback is being used to load multiple tests
    iframe.src = 'support/set-cookie.php?'+encodeURIComponent('ws_test=; Path=/; HttpOnly; Expires=Sun, 06 Nov 1994 08:49:37 GMT');
  });
  document.body.appendChild(iframe);
}, null, {timeout:10000})
</script>
